Simplify telemetry utils responsibilities
Remove dead base64 cache helpers from olive/telemetry/utils.py and move exception formatting next to the telemetry extension code that actually uses it. Keep the Win32 locking and cache-dir logic intact while reducing unrelated utility clutter.

Files changed:
- olive/telemetry/utils.py
- olive/telemetry/telemetry_extensions.py
